refactor: Game에서의 게임 실행 책임을 MinesweeperGame에 위임
