+++ src/main/kotlin/minesweeper/MinesweeperGame.kt
@@ -3,6 +3,10 @@ package minesweeper
 import minesweeper.board.BoardElement
 import minesweeper.board.GameBoard
 import minesweeper.position.Position
+import minesweeper.view.Input
+import minesweeper.view.Output
+import java.lang.NumberFormatException
+import java.lang.RuntimeException
 
 class MinesweeperGame (
     private val defaultGameBoard: GameBoard,
@@ -11,7 +15,7 @@ class MinesweeperGame (
 ) {
     private val positionQueue = ArrayDeque<Position>()
 
-    fun play(position: Position): PlayStatus {
+    private fun openCell(position: Position): PlayStatus {
         if (minesweeperGameBoard.isMine(position)) {
             return PlayStatus.LOSE
         }
@@ -34,7 +38,7 @@ class MinesweeperGame (
         return if(defaultGameBoard.areAllCellsOpened()) PlayStatus.WIN else PlayStatus.OPEN
     }
 
-    fun render() = defaultGameBoard.render()
+    private fun render() = defaultGameBoard.render()
 
     private fun openNearCell(positions: List<Position>) {
         positions.forEach {
@@ -44,5 +48,48 @@ class MinesweeperGame (
                 positionQueue.add(it)
             }
         }
+    }
+
+    tailrec fun gameStart() {
+        val position = convertStringToPosition(boardElement)
+        when (this.openCell(position)) {
+            PlayStatus.OPEN -> {
+                Output.printAny(this.render())
+                gameStart()
+            }
+            PlayStatus.WIN -> Output.printWinGame()
+            PlayStatus.LOSE -> Output.printLoseGame()
+        }
+    }
+
+    private tailrec fun convertStringToPosition(
+        boardElement: BoardElement
+    ): Position {
+        Output.printCellMessage()
+        return toPosition(Input.getLine().split(INPUT_POSITION_DELIMITER), boardElement)
+            ?: convertStringToPosition(boardElement)
+    }
+
+    private fun toPosition(split: List<String>, boardElement: BoardElement): Position? {
+        try {
+            require(split.size == INPUT_SIZE)
+            val position = Position(split[COL], split[ROW])
+            if (boardElement.isOutOfRange(position)) {
+                return null
+            }
+            return position
+        } catch (e: RuntimeException) {
+            return when(e) {
+                is IllegalArgumentException, is NumberFormatException -> null
+                else -> throw e
+            }
+        }
+    }
+
+    companion object {
+        private const val ROW = 0
+        private const val COL = 1
+        private const val INPUT_SIZE = 2
+        private const val INPUT_POSITION_DELIMITER = ", "
     }
 }
